requirements.txt
_tflow_select==2.3.0
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
ale-py==0.8.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
arviz==0.11.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
atari-py==1.2.2
atomicwrites==1.4.1
attrs==22.1.0
babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
blas==1.0
bleach==4.1.0
blinker==1.5
boost-cpp==1.78.0
boost==1.78.0
bottleneck==1.3.5
box2d-py==2.3.5
box2d==2.3.2
brotli-bin==1.0.9
brotli==1.0.9
brotlipy==0.7.0
bzip2==1.0.8
ca-certificates==2022.12.7
cachetools==5.3.0
celluloid==0.2.0
certifi==2022.12.7
cffi==1.15.1
cftime==1.5.1.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
contourpy==1.0.5
cryptography==38.0.4
cuda-cccl==11.7.91
cuda-command-line-tools==11.7.1
cuda-compiler==11.7.1
cuda-cudart-dev==11.7.99
cuda-cudart==11.7.99
cuda-cuobjdump==11.7.91
cuda-cupti==11.7.101
cuda-cuxxfilt==11.7.91
cuda-demo-suite==12.0.76
cuda-documentation==12.0.76
cuda-libraries-dev==11.7.1
cuda-libraries==11.7.1
cuda-memcheck==11.8.86
cuda-nsight-compute==12.0.0
cuda-nvcc==11.7.99
cuda-nvdisasm==12.0.76
cuda-nvml-dev==11.7.91
cuda-nvprof==12.0.90
cuda-nvprune==11.7.91
cuda-nvrtc-dev==11.7.99
cuda-nvrtc==11.7.99
cuda-nvtx==11.7.91
cuda-nvvp==12.0.90
cuda-runtime==11.7.1
cuda-sanitizer-api==12.0.90
cuda-toolkit==11.7.1
cuda-tools==11.7.1
cuda-visual-tools==11.7.1
cuda==11.7.1
cudatoolkit==11.3.1
curl==7.86.0
cycler==0.11.0
cython==0.29.32
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fasteners==0.18
flatbuffers==22.12.06
flit-core==3.6.0
fonttools==4.25.0
freetype==2.12.1
frozenlist==1.3.3
gast==0.4.0
giflib==5.2.1
git==2.34.1
glfw==2.5.5
glib==2.69.1
google-auth-oauthlib==0.4.6
google-auth==2.16.0
google-pasta==0.2.0
grpcio==1.42.0
gst-plugins-base==1.18.5
gstreamer==1.18.5
gym-notices==0.0.8
gym==0.26.2
h5py==2.10.0
hdf4==4.2.13
hdf5==1.10.6
icc_rt==2022.1.0
icu==58.2
idna==3.4
imageio-ffmpeg==0.4.7
imageio==2.22.3
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
intel-openmp==2021.4.0
ipykernel==6.9.1
ipython==8.7.0
ipython_genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.1
jinja2==3.1.2
joblib==1.1.1
jpeg==9e
jsanimation==0.1
json5==0.9.6
jsonschema==4.16.0
jupyter==1.0.0
jupyter_client==7.3.4
jupyter_console==6.4.4
jupyter_core==5.1.1
jupyter_server==1.23.4
jupyterlab-widgets==3.0.5
jupyterlab==3.5.2
jupyterlab_pygments==0.1.2
jupyterlab_server==2.16.5
jupyterlab_widgets==1.0.0
keras-preprocessing==1.1.2
keras==2.10.0
kiwisolver==1.4.4
lerc==3.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libclang==12.0.0
libcublas-dev==11.10.3.66
libcublas==11.10.3.66
libcufft-dev==10.7.2.124
libcufft==10.7.2.124
libcurand-dev==10.3.1.50
libcurand==10.3.1.50
libcurl==7.86.0
libcusolver-dev==11.4.0.1
libcusolver==11.4.0.1
libcusparse-dev==11.7.4.91
libcusparse==11.7.4.91
libdeflate==1.8
libffi==3.4.2
libiconv==1.16
libnetcdf==4.8.1
libnpp-dev==11.7.4.75
libnpp==11.7.4.75
libnvjpeg-dev==11.8.0.2
libnvjpeg==11.8.0.2
libogg==1.3.5
libpng==1.6.37
libprotobuf==3.20.3
libpython==2.1
libsodium==1.0.18
libssh2==1.10.0
libtiff==4.5.0
libuv==1.40.0
libvorbis==1.3.7
libwebp-base==1.2.4
libwebp==1.2.4
libxml2==2.9.14
libxslt==1.1.35
libzip==1.8.0
lxml==4.9.1
lz4-c==1.9.4
lz4==4.0.2
m2w64-binutils==2.25.1
m2w64-bzip2==1.0.6
m2w64-crt-git==5.0.0.4636.2595836
m2w64-gcc-ada==5.3.0
m2w64-gcc-fortran==5.3.0
m2w64-gcc-libgfortran==5.3.0
m2w64-gcc-libs-core==5.3.0
m2w64-gcc-libs==5.3.0
m2w64-gcc-objc==5.3.0
m2w64-gcc==5.3.0
m2w64-gmp==6.1.0
m2w64-headers-git==5.0.0.4636.c0ad18a
m2w64-isl==0.16.1
m2w64-libiconv==1.14
m2w64-libmangle-git==5.0.0.4509.2e5a9a2
m2w64-libwinpthread-git==5.0.0.4634.697f757
m2w64-make==4.1.2351.a80a8b8
m2w64-mpc==1.0.3
m2w64-mpfr==3.1.4
m2w64-pkg-config==0.29.1
m2w64-toolchain==5.3.0
m2w64-tools-git==5.0.0.4592.90b8472
m2w64-windows-default-manifest==6.4
m2w64-winpthreads-git==5.0.0.4634.697f757
m2w64-zlib==1.2.8
mako==1.2.3
markdown==3.4.1
markupsafe==2.1.1
matplotlib-base==3.6.2
matplotlib-inline==0.1.6
matplotlib==3.6.2
mistune==0.8.4
mkl-service==2.4.0
mkl==2021.4.0
mkl_fft==1.3.1
mkl_random==1.2.2
moviepy==1.0.3
msys2-conda-epoch==20160418
mujoco-py==2.1.2.14
mujoco==2.2.0
multidict==6.0.4
munkres==1.1.4
nb_conda_kernels==2.3.1
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
netcdf4==1.6.2
notebook-shim==0.2.2
notebook==6.5.2
nsight-compute==2022.4.0.15
numexpr==2.8.4
numpy-base==1.21.5
numpy==1.23.4
nvidia-cuda-runtime-cu12==12.0.107
nvidia-pyindex==1.0.9
oauthlib==3.2.2
opencv-python==4.6.0.66
openssl==1.1.1t
opt_einsum==3.3.0
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pcre==8.45
pickleshare==0.7.5
pillow==9.3.0
pip==22.2.2
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
proglog==0.1.10
prometheus_client==0.14.1
prompt-toolkit==3.0.36
prompt_toolkit==3.0.36
protobuf==3.20.3
pure_eval==0.2.2
py==1.11.0
pyasn1-modules==0.2.7
pyasn1==0.4.8
pycparser==2.21
pycuda==2022.2.2
pygame==2.1.0
pyglet==1.2.4
pygments==2.11.2
pyjwt==2.6.0
pyopengl==3.1.6
pyopenssl==22.0.0
pyparsing==3.0.9
pyqt5-sip==12.11.0
pyqt==5.15.7
pyreadline==2.1
pyrsistent==0.18.0
pysocks==1.7.1
pystan==2.19.1.1
pytest==7.0.1
python-dateutil==2.8.2
python-fastjsonschema==2.16.2
python-flatbuffers==23.1.21
python==3.9.13
python_abi==3.9
pytools==2022.1.14
pytorch-cuda==11.7
pytorch-mutex==1.0
pytorch==1.13.1
pytz==2022.7
pyu2f==0.1.5
pyvirtualdisplay==3.0
pywin32==305
pywinpty==2.0.2
pyzmq==23.2.0
qt-main==5.15.2
qt-webengine==5.15.9
qtconsole==5.3.2
qtpy==2.2.0
qtwebkit==5.212
requests-oauthlib==1.3.1
requests==2.28.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.7.3
seaborn==0.11.2
send2trash==1.8.0
setuptools==65.6.3
sip==6.6.2
six==1.16.0
snappy==1.1.9
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlite==3.40.1
stack_data==0.2.0
swig==4.1.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.10.0
tensorflow-base==2.10.0
tensorflow-estimator==2.10.0
tensorflow==2.10.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==2.2.0
tinycss2==1.2.1
tk==8.6.12
toml==0.10.2
tomli==2.0.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.1
tqdm==4.64.1
traitlets==5.7.1
typing-extensions==4.4.0
typing_extensions==4.4.0
tzdata==2022g
ucrt==10.0.20348.0
urllib3==1.26.12
vc==14.2
vs2015_runtime==14.34.31931
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==4.0.5
win_inet_pton==1.1.0
wincertstore==0.2
winpty==0.4.3
wrapt==1.14.1
xarray==2022.11.0
xz==5.2.8
yarl==1.8.2
zeromq==4.3.4
zipp==3.10.0
zlib==1.2.13
zstd==1.5.2
